requirements.txt
### NSFR
gym
gym3
#torch==1.12.1
lark-parser
sklearn
gym
gym3
filelock
pyglet
torchsummary
setuptools>=58.5.3
tensorboard
numpy
#torchvision
tqdm>=4.41.0
matplotlib>=3.2.2
opencv-python>=4.1.2
Pillow
PyYAML>=5.3.1
scipy>=1.4.1
seaborn
pandas
cmake
rtpt
wandb
pyglet
###qt
